Adjust log APIs.
