--- apm-commons/apm-logging-api/src/main/java/org/skywalking/apm/logging/LogResolver.java
@@ -1,32 +1,0 @@
-/*
- * Copyright 2017, OpenSkywalking Organization All rights reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- * Project repository: https://github.com/OpenSkywalking/skywalking
- */
-
-package org.skywalking.apm.logging;
-
-/**
- * {@link LogResolver} just do only one thing: return the {@link ILog} implementation.
- * <p>
- * Created by xin on 2016/11/10.
- */
-public interface LogResolver {
-    /**
-     * @param clazz, the class is showed in log message.
-     * @return {@link ILog} implementation.
-     */
-    ILog getLogger(Class<?> clazz);
-}
