+++ apm-sniffer/apm-agent-core/src/main/java/org/skywalking/apm/agent/core/logging/core/SystemOutWriter.java
@@ -16,17 +16,22 @@
  * Project repository: https://github.com/OpenSkywalking/skywalking
  */
 
-package org.skywalking.apm.agent.core.logging;
+package org.skywalking.apm.agent.core.logging.core;
 
-import org.skywalking.apm.logging.ILog;
-import org.skywalking.apm.logging.LogResolver;
+import java.io.PrintStream;
 
-/**
- * Created by wusheng on 2016/11/26.
- */
-public class EasyLogResolver implements LogResolver {
+public enum SystemOutWriter implements IWriter {
+    INSTANCE;
+
+    /**
+     * Tricky codes for avoiding style-check.
+     * Because, in here, "system.out.println" is the only choice to output logs.
+     *
+     * @param message
+     */
     @Override
-    public ILog getLogger(Class<?> clazz) {
-        return new EasyLogger(clazz);
+    public void write(String message) {
+        PrintStream out = System.out;
+        out.println(message);
     }
 }
