+++ apm-sniffer/apm-agent-core/src/main/java/org/skywalking/apm/agent/core/plugin/PluginBootstrap.java
@@ -23,8 +23,8 @@ import java.util.ArrayList;
 import java.util.List;
 import org.skywalking.apm.agent.core.boot.AgentPackageNotFoundException;
 import org.skywalking.apm.agent.core.plugin.loader.AgentClassLoader;
-import org.skywalking.apm.logging.ILog;
-import org.skywalking.apm.logging.LogManager;
+import org.skywalking.apm.agent.core.logging.api.ILog;
+import org.skywalking.apm.agent.core.logging.api.LogManager;
 
 /**
  * Plugins finder.
