+++ apm-sniffer/apm-agent-core/src/main/java/org/skywalking/apm/agent/core/plugin/PluginResourcesResolver.java
@@ -24,8 +24,8 @@ import java.util.ArrayList;
 import java.util.Enumeration;
 import java.util.List;
 import org.skywalking.apm.agent.core.plugin.loader.AgentClassLoader;
-import org.skywalking.apm.logging.ILog;
-import org.skywalking.apm.logging.LogManager;
+import org.skywalking.apm.agent.core.logging.api.ILog;
+import org.skywalking.apm.agent.core.logging.api.LogManager;
 
 /**
  * Use the current classloader to read all plugin define file.
