+++ apm-sniffer/apm-agent-core/src/main/java/org/skywalking/apm/agent/core/plugin/loader/AgentClassLoader.java
@@ -21,8 +21,8 @@ package org.skywalking.apm.agent.core.plugin.loader;
 import org.skywalking.apm.agent.core.boot.AgentPackageNotFoundException;
 import org.skywalking.apm.agent.core.boot.AgentPackagePath;
 import org.skywalking.apm.agent.core.plugin.PluginBootstrap;
-import org.skywalking.apm.logging.ILog;
-import org.skywalking.apm.logging.LogManager;
+import org.skywalking.apm.agent.core.logging.api.ILog;
+import org.skywalking.apm.agent.core.logging.api.LogManager;
 
 import java.io.*;
 import java.net.MalformedURLException;
