+++ apm-sniffer/apm-agent-core/src/main/java/org/skywalking/apm/agent/core/remote/TraceSegmentServiceClient.java
@@ -29,8 +29,8 @@ import org.skywalking.apm.agent.core.context.trace.TraceSegment;
 import org.skywalking.apm.commons.datacarrier.DataCarrier;
 import org.skywalking.apm.commons.datacarrier.buffer.BufferStrategy;
 import org.skywalking.apm.commons.datacarrier.consumer.IConsumer;
-import org.skywalking.apm.logging.ILog;
-import org.skywalking.apm.logging.LogManager;
+import org.skywalking.apm.agent.core.logging.api.ILog;
+import org.skywalking.apm.agent.core.logging.api.LogManager;
 import org.skywalking.apm.network.proto.Downstream;
 import org.skywalking.apm.network.proto.TraceSegmentServiceGrpc;
 import org.skywalking.apm.network.proto.UpstreamSegment;
