+++ apm-sniffer/apm-agent-core/src/test/java/org/skywalking/apm/agent/core/logging/core/EasyLoggerTest.java
@@ -16,15 +16,16 @@
  * Project repository: https://github.com/OpenSkywalking/skywalking
  */
 
-package org.skywalking.apm.agent.core.logging;
+package org.skywalking.apm.agent.core.logging.core;
 
-import java.io.PrintStream;
 import org.junit.AfterClass;
 import org.junit.Assert;
 import org.junit.BeforeClass;
 import org.junit.Test;
 import org.mockito.Mockito;
 import org.skywalking.apm.agent.core.conf.Constants;
+
+import java.io.PrintStream;
 
 import static org.mockito.Matchers.anyString;
 import static org.mockito.Mockito.times;
@@ -48,7 +49,12 @@ public class EasyLoggerTest {
         System.setOut(output);
         PrintStream err = Mockito.mock(PrintStream.class);
         System.setErr(err);
-        EasyLogger logger = new EasyLogger(EasyLoggerTest.class);
+        EasyLogger logger = new EasyLogger(EasyLoggerTest.class) {
+            @Override
+            protected void logger(LogLevel level, String message, Throwable e) {
+                SystemOutWriter.INSTANCE.write(format(level, message, e));
+            }
+        };
 
         Assert.assertTrue(logger.isDebugEnable());
         Assert.assertTrue(logger.isInfoEnable());
@@ -67,7 +73,7 @@ public class EasyLoggerTest {
         logger.error(new NullPointerException(), "hello {}", "world");
 
         Mockito.verify(output, times(9))
-            .println(anyString());
+                .println(anyString());
     }
 
     @Test
@@ -76,7 +82,12 @@ public class EasyLoggerTest {
         System.setOut(output);
         PrintStream err = Mockito.mock(PrintStream.class);
         System.setErr(err);
-        EasyLogger logger = new EasyLogger(EasyLoggerTest.class);
+        EasyLogger logger = new EasyLogger(EasyLoggerTest.class) {
+            @Override
+            protected void logger(LogLevel level, String message, Throwable e) {
+                SystemOutWriter.INSTANCE.write(format(level, message, e));
+            }
+        };
 
         Assert.assertTrue(logger.isDebugEnable());
         Assert.assertTrue(logger.isInfoEnable());
@@ -95,7 +106,7 @@ public class EasyLoggerTest {
         logger.error(new NullPointerException(), "hello {}", "&&&**%%");
 
         Mockito.verify(output, times(9))
-            .println(anyString());
+                .println(anyString());
     }
 
     @Test
@@ -105,7 +116,7 @@ public class EasyLoggerTest {
         String formatLines = logger.format(exception);
         String[] lines = formatLines.split(Constants.LINE_SEPARATOR);
         Assert.assertEquals("java.lang.NullPointerException", lines[1]);
-        Assert.assertEquals("\tat org.skywalking.apm.agent.core.logging.EasyLoggerTest.testFormat(EasyLoggerTest.java:103)", lines[2]);
+        Assert.assertEquals("\tat org.skywalking.apm.agent.core.logging.core.EasyLoggerTest.testFormat(EasyLoggerTest.java:114)", lines[2]);
     }
 
     @AfterClass
