+++ apm-sniffer/apm-agent/src/main/java/org/skywalking/apm/agent/SkyWalkingAgent.java
@@ -18,23 +18,18 @@
 
 package org.skywalking.apm.agent;
 
-import java.lang.instrument.Instrumentation;
-import java.util.List;
 import net.bytebuddy.agent.builder.AgentBuilder;
 import net.bytebuddy.description.type.TypeDescription;
 import net.bytebuddy.dynamic.DynamicType;
 import net.bytebuddy.utility.JavaModule;
 import org.skywalking.apm.agent.core.boot.ServiceManager;
 import org.skywalking.apm.agent.core.conf.SnifferConfigInitializer;
-import org.skywalking.apm.agent.core.logging.EasyLogResolver;
-import org.skywalking.apm.agent.core.logging.SystemOutWriter;
-import org.skywalking.apm.agent.core.plugin.AbstractClassEnhancePluginDefine;
-import org.skywalking.apm.agent.core.plugin.EnhanceContext;
-import org.skywalking.apm.agent.core.plugin.PluginBootstrap;
-import org.skywalking.apm.agent.core.plugin.PluginException;
-import org.skywalking.apm.agent.core.plugin.PluginFinder;
-import org.skywalking.apm.logging.ILog;
-import org.skywalking.apm.logging.LogManager;
+import org.skywalking.apm.agent.core.logging.api.ILog;
+import org.skywalking.apm.agent.core.logging.api.LogManager;
+import org.skywalking.apm.agent.core.plugin.*;
+
+import java.lang.instrument.Instrumentation;
+import java.util.List;
 
 /**
  * The main entrance of sky-waking agent,
@@ -43,12 +38,7 @@ import org.skywalking.apm.logging.LogManager;
  * @author wusheng
  */
 public class SkyWalkingAgent {
-    private static final ILog logger;
-
-    static {
-        LogManager.setLogResolver(new EasyLogResolver());
-        logger = LogManager.getLogger(SkyWalkingAgent.class);
-    }
+    private static final ILog logger = LogManager.getLogger(SkyWalkingAgent.class);
 
     /**
      * Main entrance.
@@ -67,8 +57,7 @@ public class SkyWalkingAgent {
 
             ServiceManager.INSTANCE.boot();
         } catch (Exception e) {
-            SystemOutWriter.INSTANCE.write("skywalking agent is shutting down.");
-            e.printStackTrace(SystemOutWriter.INSTANCE.getStream());
+            logger.error(e, "Skywalking agent initialized failure. Shutting down.");
             return;
         }
